+++ next-memmem/prisma/schema.prisma
@@ -117,6 +117,7 @@ model tbl_user {
   tbl_company tbl_company[]
   tbl_notice  tbl_notice[]
   tbl_role    tbl_role[]    @ignore
+  tbl_user_comp tbl_user_comp[]
 }
 
 model tbl_user_comp {
@@ -128,6 +129,7 @@ model tbl_user_comp {
   us_date  String? @db.VarChar(15)
 
   @@id([us_uid, us_ccode])
+  tbl_user tbl_user @relation(fields: [us_uid], references: [u_id])
 }
 
 model tbl_user_minfo {
